# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type SchoolworkKind = typeof SchoolworkKind[keyof typeof SchoolworkKind];


export const SchoolworkKind = {
  worksheet_problem: 'worksheet_problem',
  assignment_project: 'assignment_project',
  concept_topic: 'concept_topic',
} as const;

export type PhotoAnalysisInputSelectedMode = typeof PhotoAnalysisInputSelectedMode[keyof typeof PhotoAnalysisInputSelectedMode];


export const PhotoAnalysisInputSelectedMode = {
  assignment: 'assignment',
  concept: 'concept',
} as const;

export interface PhotoAnalysisInput {
  /** @minLength 20 */
  imageDataUrl: string;
  studentRequest: string;
  selectedMode: PhotoAnalysisInputSelectedMode;
}

export interface PhotoExtraction {
  kind: SchoolworkKind;
  title: string;
  directions: string;
  visibleContent: string;
  gradeLevel: string;
  subject: string;
  skill: string;
  requirements: string[];
  readable: boolean;
  note: string;
}

export interface PhotoHelpInput {
  extraction: PhotoExtraction;
  studentRequest: string;
}

export type PhotoHelpResultActualProblemsItem = {
  problem: string;
  hints: string[];
};

export type PhotoHelpResultProjectIdeasItem = {
  title: string;
  description: string;
  approach: string;
  whyItFits: string;
};

export type PhotoHelpResultDetectedCitationStyle = typeof PhotoHelpResultDetectedCitationStyle[keyof typeof PhotoHelpResultDetectedCitationStyle];


export const PhotoHelpResultDetectedCitationStyle = {
  MLA: 'MLA',
  APA: 'APA',
  Chicago: 'Chicago',
  Not_specified: 'Not specified',
} as const;

export type ResearchResourceResourceType = typeof ResearchResourceResourceType[keyof typeof ResearchResourceResourceType];


export const ResearchResourceResourceType = {
  article: 'article',
  activity: 'activity',
  image: 'image',
  video: 'video',
} as const;

export interface ResearchResource {
  title: string;
  organization: string;
  credibility: string;
  supports: string;
  date: string;
  url: string;
  resourceType: ResearchResourceResourceType;
}

export interface PhotoHelpResult {
  kind: SchoolworkKind;
  heading: string;
  explanation: string[];
  keyIdeas: string[];
  exampleProblem: string;
  exampleSteps: string[];
  exampleAnswer: string;
  actualProblems: PhotoHelpResultActualProblemsItem[];
  guidedTry: string;
  understandingCheck: string;
  summary: string;
  planSteps: string[];
  projectIdeas: PhotoHelpResultProjectIdeasItem[];
  resources: ResearchResource[];
  detectedCitationStyle: PhotoHelpResultDetectedCitationStyle;
  /** Explicit items the student must submit, present, bring, upload, or hand in */
  deliverables: string[];
  /** Explicit submission method or Turn-in method not provided */
  turnInMethod: string;
}

export type ProjectPlanInputSelectedIdea = {
  title: string;
  description: string;
  approach: string;
  whyItFits: string;
};

export type ProjectPlanInputCitationStyle = typeof ProjectPlanInputCitationStyle[keyof typeof ProjectPlanInputCitationStyle];


export const ProjectPlanInputCitationStyle = {
  MLA: 'MLA',
  APA: 'APA',
  Chicago: 'Chicago',
} as const;

export interface ProjectPlanInput {
  extraction: PhotoExtraction;
  selectedIdea: ProjectPlanInputSelectedIdea;
  citationStyle: ProjectPlanInputCitationStyle;
}

export interface ProjectPlanResult {
  summary: string;
  planSteps: string[];
}

export interface ConceptHelpInput {
  /**
     * @minLength 3
     * @maxLength 2000
     */
  studentRequest: string;
}

export interface ConceptHelpResult {
  topic: string;
  gradeLevel: string;
  subject: string;
  studentGoal: string;
  heading: string;
  /** @minItems 1 */
  explanation: string[];
  /** @minItems 2 */
  keyIdeas: string[];
  exampleProblem: string;
  /** @minItems 2 */
  exampleSteps: string[];
  exampleAnswer: string;
  guidedTry: string;
  practiceProblem: string;
  /** @minItems 2 */
  practiceHints: string[];
  /** @minItems 2 */
  practiceSteps: string[];
  understandingCheck: string;
}

export interface ConceptAnswerCheckInput {
  /**
     * @minLength 3
     * @maxLength 2000
     */
  studentRequest: string;
  topic: string;
  gradeLevel: string;
  subject: string;
  practiceProblem: string;
  context: string;
  /**
     * @minLength 1
     * @maxLength 1000
     */
  studentAnswer: string;
}

export interface ConceptAnswerCheckResult {
  isCorrect: boolean;
  feedback: string;
}

export type ProblemAnswerRevealInputFlow = typeof ProblemAnswerRevealInputFlow[keyof typeof ProblemAnswerRevealInputFlow];


export const ProblemAnswerRevealInputFlow = {
  concept: 'concept',
  photo: 'photo',
} as const;

export interface ProblemAnswerRevealInput {
  flow: ProblemAnswerRevealInputFlow;
  /** @maxLength 2000 */
  studentRequest: string;
  topic: string;
  gradeLevel: string;
  subject: string;
  problem: string;
  /** @maxLength 10000 */
  context: string;
}

export interface ProblemAnswerRevealResult {
  /** @minItems 1 */
  reasoning: string[];
  answer: string;
}