# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /concept-help:
    post:
      operationId: createConceptHelp
      tags: [concept-help]
      summary: Generate an adaptive concept lesson from a student request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConceptHelpInput"
      responses:
        "200":
          description: Parsed intent and grade-appropriate teaching sequence
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConceptHelpResult"
        "400":
          description: Invalid or too-short request
        "502":
          description: AI provider failed
  /concept-answer-check:
    post:
      operationId: checkConceptAnswer
      tags: [concept-help]
      summary: Check a student's concept-practice answer without revealing the solution
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConceptAnswerCheckInput"
      responses:
        "200":
          description: Semantic correctness and student-safe feedback
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConceptAnswerCheckResult"
        "400":
          description: Invalid answer-check request
        "502":
          description: AI provider failed
  /problem-answer-reveal:
    post:
      operationId: revealProblemAnswer
      tags: [concept-help, photo-help]
      summary: Reveal reasoning and a final answer after explicit student request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProblemAnswerRevealInput"
      responses:
        "200":
          description: Step-by-step reasoning and final answer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemAnswerRevealResult"
        "400":
          description: Invalid reveal request
        "502":
          description: AI provider failed
components:
  schemas:
...
    ConceptHelpInput:
      type: object
      properties:
        studentRequest:
          type: string
          minLength: 3
          maxLength: 2000
      required: [studentRequest]
    ConceptHelpResult:
      type: object
      properties:
        topic:
          type: string
        gradeLevel:
          type: string
        subject:
          type: string
        studentGoal:
          type: string
        heading:
          type: string
        explanation:
          type: array
          minItems: 1
          items:
            type: string
        keyIdeas:
          type: array
          minItems: 2
          items:
            type: string
        exampleProblem:
          type: string
        exampleSteps:
          type: array
          minItems: 2
          items:
            type: string
        exampleAnswer:
          type: string
        guidedTry:
          type: string
        practiceProblem:
          type: string
        practiceHints:
          type: array
          minItems: 2
          items:
            type: string
        practiceSteps:
          type: array
          minItems: 2
          items:
            type: string
        understandingCheck:
          type: string
      required: [topic, gradeLevel, subject, studentGoal, heading, explanation, keyIdeas, exampleProblem, exampleSteps, exampleAnswer, guidedTry, practiceProblem, practiceHints, practiceSteps, understandingCheck]
    ConceptAnswerCheckInput:
      type: object
      properties:
        studentRequest:
          type: string
          minLength: 3
          maxLength: 2000
        topic:
          type: string
        gradeLevel:
          type: string
        subject:
          type: string
        practiceProblem:
          type: string
        context:
          type: string
        studentAnswer:
          type: string
          minLength: 1
          maxLength: 1000
      required: [studentRequest, topic, gradeLevel, subject, practiceProblem, context, studentAnswer]
    ConceptAnswerCheckResult:
      type: object
      properties:
        isCorrect:
          type: boolean
        feedback:
          type: string
      required: [isCorrect, feedback]
    ProblemAnswerRevealInput:
      type: object
      properties:
        flow:
          type: string
          enum: [concept, photo]
        studentRequest:
          type: string
          maxLength: 2000
        topic:
          type: string
        gradeLevel:
          type: string
        subject:
          type: string
        problem:
          type: string
        context:
          type: string
          maxLength: 10000
      required: [flow, studentRequest, topic, gradeLevel, subject, problem, context]
    ProblemAnswerRevealResult:
      type: object
      properties:
        reasoning:
          type: array
          minItems: 1
          items:
            type: string
        answer:
          type: string
      required: [reasoning, answer]