# lib/api-zod/src/generated/types/conceptAnswerCheckInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ConceptAnswerCheckInput {
  /**
     * @minLength 3
     * @maxLength 2000
     */
  studentRequest: string;
  topic: string;
  gradeLevel: string;
  subject: string;
  practiceProblem: string;
  context: string;
  /**
     * @minLength 1
     * @maxLength 1000
     */
  studentAnswer: string;
}