# lib/api-zod/src/generated/types/conceptHelpResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ConceptHelpResult {
  topic: string;
  gradeLevel: string;
  subject: string;
  studentGoal: string;
  heading: string;
  /** @minItems 1 */
  explanation: string[];
  /** @minItems 2 */
  keyIdeas: string[];
  exampleProblem: string;
  /** @minItems 2 */
  exampleSteps: string[];
  exampleAnswer: string;
  guidedTry: string;
  practiceProblem: string;
  /** @minItems 2 */
  practiceHints: string[];
  /** @minItems 2 */
  practiceSteps: string[];
  understandingCheck: string;
}