# lib/api-zod/src/generated/types/problemAnswerRevealInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProblemAnswerRevealInputFlow } from './problemAnswerRevealInputFlow';

export interface ProblemAnswerRevealInput {
  flow: ProblemAnswerRevealInputFlow;
  /** @maxLength 2000 */
  studentRequest: string;
  topic: string;
  gradeLevel: string;
  subject: string;
  problem: string;
  /** @maxLength 10000 */
  context: string;
}