# lib/api-zod/src/generated/types/problemAnswerRevealResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProblemAnswerRevealResult {
  /** @minItems 1 */
  reasoning: string[];
  answer: string;
}